[configure.in] Use AC_MSG_RESULT in PKG_CHECK_MODULES's second branch
If PKG_CHECK_MODULES fails, it does not print out any check results and so, no
newlines.  This is kinda silly, at least in the case that no second branch is
provided, but I think that's the way it is, to let users decide what to print.

We now just do a AC_MSG_RESULT(no) and continue with what we used to do.
Inspired by Tor Lillqvist's similar change in Pango.