Expand font_map locking to cover call to backend->scaled_font_create
It seemed like a good idea to avoid holding the lock over the call
into the backend. But the procedure in place was quite broken:

	LOCK
	Fail to find font in hash table
	UNLOCK
	Create font
	LOCK
	Insert font into hash table
	UNLOCK

since while we're busy creating the font and unlocked, another thread
can insert the font. Our paranoid hash table implementation noted
the problem quite readily.

This patch simply removes the internal UNLOCK/LOCK to hold the mutex
over the whole procedure.