[PS] Update surface_pattern_supported() from PDF
The PS backend was accepting all surface patterns, just to ASSERT_NOT_REACHED
later if extend type of the pattern was PAD or REFLECT.  That didn't make
sense and was causing crash for PAD (surprisingly not for REFLECT.  Not sure
why).  So now it rejects those types of surface patterns, just like the PDF
surface does.