Revert the solid-pattern cache
This reverts the following commits:

	2715f2098127d04d2f9e304580a26cd0842c0e64
	67e3b3c53bdd69e4d3503eed2db66953f1ac038c

See this thread for an analysis of the problems it caused:

	http://lists.freedesktop.org/archives/cairo/2007-February/009825.html

In short, a single cache for all backends doesn't work, as one thread
using any backend can cause an unused xlib pattern to be evicted from
the cache, and trigger an xlib call while the display is being used
from another thread.  Xlib is not prepared for this.