Implement CAIRO_MUTEX_INIT with memcpy instead of pthread_mutex_init
The trick here is that with the weak symbol support for pthreads,
pthread_mutex_init can be a NOP leaving the mutex uninitialized.
Then, if some pthread-using library is dynamically loaded, the
non-NOP pthread functions get used and we end up trying to lock
an uninitialized mutex.

This should fix the bugs reported here:

	Cairo 1.3.14 deadlocks in cairo_scaled_font_glyph_extents or _cairo_ft_unscaled_font_lock_face
	https://bugs.freedesktop.org/show_bug.cgi?id=10035