Mark 5 tests that are currently failing as XFAIL (that is, not to be fixed before 1.4)
The 5 additional bugs that will be shipped with 1.4 are

	ft-text-vertical-layout-type1
	radial-gradient
	surface-pattern
	surface-pattern-scale-down
	surface-pattern-scale-up

Most of these are non-issues, (unbundled font for
ft-text-vertical-layout-type1), or very minor issues (radial-gradient
and surface-pattern). The only things in here that look like a real
bug are the surface-pattern-scale-down and surface-pattern-scale-up
tests where the xlib backend results have some non-1.0 alpha that is
very unexpected.