Disable warnings for cairo_*_reference functions.
This class of functions modify their argument and return it as a
*convenience* for the caller. For example, within cairo a common idiom is:

    cairo_object_reference (new);
    cairo_object_unreference (this->obj);
    this->obj = new;

which updates the member to the new object irrespective of whether the
new object is the same as the one being replaced. Other issues arise
with subtypes, as the return type is the parent's and so require more
complicated handling to compile cleanly.

Disabling the warning is therefore preferred over adding code which
decreases readibility and reduces maintainability. We need to make the
compiler work for us, not against us...