test/invalid-matrix: Add new test to exercise CAIRO_STATUS_INVALID_MATRIX paths
This new test exercises every path where the user might possibly
pass in an invalid matrix. Currently the test fails if no error
is reported. Also, if an incorrect error is reported, (such as
CAIRO_STATUS_NO_MEMORY instead of CAIRO_STATUS_INVALID_MATRIX),
this is logged as a warning in invalid-matrix.log, but the test
still passes.

It would still be worthwhile to follow up quickly and fix those
cases to propagate the correct error value.