Restore the error message for unsupported X server visuals
Apparently, our much-beloved error message, (just look at how many people
love to quote it in bugzilla entries), was removed during the switch to
the new pixman. It was replaced with an obscure failure of a nil surface
being returned, (leading to obscure failures or false "out of memory"
errors).

So we're putting this back now, (even though we _really_ want to ensure
that this message never gets printed in practice).